pyproject.toml
[tool.poetry]
name = "rabbitwebsocketchat"
version = "0.1.0"
description = ""
authors = ["Igor <iiigorrr@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.5"}
faststream = {extras = ["rebbitmq"], version = "^0.5.30"}
redis = "^5.2.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = {extras = ["postgresql"], version = "^2.0.36"}
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aio-pika = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
